code now compiles. todo: initfile

+++ src/MTL4Types.h
@@ -5,11 +5,12 @@
 #include <boost/numeric/mtl/mtl.hpp>
 
 // TODO: find replacement for this explicit typedefs
-namespace AMDiS {
-
-  namespace MTLTypes {
+namespace AMDiS 
+{
+  namespace MTLTypes 
+  {
     typedef double value_type;
-    typedef unsigned int size_type;
+    typedef unsigned long size_type;
     typedef mtl::matrix::parameters<mtl::row_major, mtl::index::c_index, mtl::non_fixed::dimensions, false, size_type> para;
     typedef mtl::matrix::compressed2D< value_type, para > MTLMatrix;
     typedef mtl::vector::dense_vector< value_type > MTLVector;
